[AMDGPU] Introduce asyncmark/wait intrinsics
Asynchronous operations are memory transfers (usually between the global memory
and LDS) that are completed independently at an unspecified scope. A thread that
requests one or more asynchronous transfers can use *async markers* to track
their completion. The thread waits for each marker to be
*completed*, which indicates that requests initiated in program order before
this marker have also completed.

For now, we implement asyncmark/wait operations on pre-GFX12 architectures
that support "LDS DMA" operations. These "legacy" operations are now extended to
accept an optional `ASYNC` parameter as a bit in the auxiliary argument. When
set, it indicates that the compiler should not automatically track the
completion of this operation.

Future work will extend support to GFX12Plus architectures that support "true"
async operations.

Co-authored-by: Ryan Mitchell <ryan.mitchell@amd.com>

Fixes: SWDEV-521121